Move test to another file
