+++ tests/integration/test_modify_engine_on_restart/configs/config.d/clusters.xml
@@ -15,8 +15,6 @@
     </cluster>
 </remote_servers>
 
-<default_replica_path>/clickhouse/tables/{database}/{table}/{uuid}</default_replica_path>
-
 <macros>
         <shard>01</shard>
 </macros>
