+++ tests/integration/test_modify_engine_on_restart/configs/config.d/clusters_zk_path.xml
@@ -15,6 +15,6 @@
         <shard>01</shard>
 </macros>
 
-<default_replica_path>/lol/kek/'/{uuid}</default_replica_path>
+<default_replica_path>/clickhouse/'/{database}/{table}/{uuid}</default_replica_path>
 
 </clickhouse>
